requirements.txt
python-telegram-bot==13.15
supabase==2.8.0
python-dotenv==1.0.1
urllib3==1.26.20
six==1.16.0